pyproject.toml
[project]
name = "card-bank"
version = "0.1.0"
description = "Проект для обработки данных банковских операций"
authors = [
    { name = "DESKTOP-8F2LDRV\\ANDRAQ", email = "mymillions@ya.ru" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["requests (>=2.32.3,<3.0.0)", "python-dotenv (>=1.0.1,<2.0.0)", "pandas (>=2.2.3,<3.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^25.1.0"
isort = "^6.0.0"
mypy = "^1.14.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^7.1.1"
black = "^25.1.0"
isort = "^6.0.0"
mypy = "^1.14.1"

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"

